--- src/main/groovy/com/github/eirnym/js2p/Js2DCExtension.java
@@ -1,4 +1,0 @@
-package com.github.eirnym.js2p;
-
-public class Js2DCExtension {
-}
