+++ src/main/groovy/com/github/js2d/Js2dExtension.java
@@ -1,0 +1,4 @@
+package com.github.js2d;
+
+public class Js2dExtension {
+}
