+++ src/test/groovy/com/github/js2d/TestCasesGenerator.groovy
@@ -1,11 +1,11 @@
-package com.github.eirnym.js2p
+package com.github.js2d
 
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
 
-import static com.github.eirnym.js2p.JsonSchemaPlugin.PLUGIN_ID
-import static com.github.eirnym.js2p.JsonSchemaPlugin.TARGET_FOLDER_BASE
+import static com.github.js2d.JsonSchemaPlugin.PLUGIN_ID
+import static com.github.js2d.JsonSchemaPlugin.TARGET_FOLDER_BASE
 
 class TestCasesGenerator {
     public static final String TARGET_FOLDER_BASE_CUSTOM = TARGET_FOLDER_BASE + 's'
